refactor: Points, Payment 디렉터리로 기능 분리
