+++ src/main/java/com/sscanner/team/payment/entity/PaymentRecord.java
@@ -1,4 +1,4 @@
-package com.sscanner.team.points.entity;
+package com.sscanner.team.payment.entity;
 
 import com.sscanner.team.User;
 import com.sscanner.team.global.common.BaseEntity;
