+++ src/main/java/com/sscanner/team/payment/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
-package com.sscanner.team.points.repository;
+package com.sscanner.team.payment.repository;
 
-import com.sscanner.team.points.entity.PaymentRecord;
+import com.sscanner.team.payment.entity.PaymentRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentRepository extends JpaRepository<PaymentRecord, String> {
