+++ src/main/java/com/sscanner/team/payment/requestdto/PaymentRequestDto.java
@@ -1,6 +1,6 @@
-package com.sscanner.team.points.requestdto;
+package com.sscanner.team.payment.requestdto;
 
-import com.sscanner.team.points.entity.PaymentRecord;
+import com.sscanner.team.payment.entity.PaymentRecord;
 import com.sscanner.team.products.entity.Product;
 import com.sscanner.team.User;
