--- src/main/java/com/sscanner/team/points/service/PaymentService.java
@@ -1,8 +1,0 @@
-package com.sscanner.team.points.service;
-
-import com.sscanner.team.points.requestdto.PointPaymentRequestDto;
-import com.sscanner.team.points.responsedto.PointResponseDto;
-
-public interface PaymentService {
-    PointResponseDto payPoint(PointPaymentRequestDto pointPaymentRequestDto);
-}
